+++ src/controller/AccountController.ts
@@ -1,4 +1,5 @@
 import { Request, Response } from "express"
+import { AccountBusiness } from "../business/AccountBusiness"
 import { AccountDatabase } from "../database/AccountDatabase"
 import { Account } from "../models/Account"
 import { AccountDB } from "../types"
@@ -6,17 +7,11 @@ import { AccountDB } from "../types"
 export class AccountController {
     public getAccounts = async (req: Request, res: Response) => {
         try {
-            const accountDatabase = new AccountDatabase()
-            const accountsDB: AccountDB[] = await accountDatabase.findAccounts()
-    
-            const accounts = accountsDB.map((accountDB) => new Account(
-                accountDB.id,
-                accountDB.balance,
-                accountDB.owner_id,
-                accountDB.created_at
-            ))
-    
-            res.status(200).send(accounts)
+            //entrada
+            const accountBusiness = new AccountBusiness()
+            const output = await accountBusiness.getAccounts()            
+          //res
+            res.status(200).send(output)
         } catch (error) {
             console.log(error)
     
@@ -35,25 +30,12 @@ export class AccountController {
     public getAccountBalance = async (req: Request, res: Response) => {
         try {
             const id = req.params.id
+
+            const accountBusiness = new AccountBusiness()
+            const output = await accountBusiness.getAccountBalance(id)
+
     
-            const accountDatabase = new AccountDatabase()
-            const accountDB = await accountDatabase.findAccountById(id)
-    
-            if (!accountDB) {
-                res.status(404)
-                throw new Error("'id' não encontrado")
-            }
-    
-            const account = new Account(
-                accountDB.id,
-                accountDB.balance,
-                accountDB.owner_id,
-                accountDB.created_at
-            )
-    
-            const balance = account.getBalance()
-    
-            res.status(200).send({ balance })
+            res.status(200).send({ output })
         } catch (error) {
             console.log(error)
     
@@ -71,43 +53,14 @@ export class AccountController {
 
     public createAccount = async (req: Request, res: Response) => {
         try {
-            const { id, ownerId } = req.body
-    
-            if (typeof id !== "string") {
-                res.status(400)
-                throw new Error("'id' deve ser string")
+            const input ={ 
+                id: req.body.id,
+                ownerId: req.body.ownerId,
             }
+            const accountBusiness = new AccountBusiness()
+            const output = await accountBusiness.createAccount(input)
     
-            if (typeof ownerId !== "string") {
-                res.status(400)
-                throw new Error("'ownerId' deve ser string")
-            }
-    
-            const accountDatabase = new AccountDatabase()
-            const accountDBExists = await accountDatabase.findAccountById(id)
-    
-            if (accountDBExists) {
-                res.status(400)
-                throw new Error("'id' já existe")
-            }
-    
-            const newAccount = new Account(
-                id,
-                0,
-                ownerId,
-                new Date().toISOString()
-            )
-    
-            const newAccountDB: AccountDB = {
-                id: newAccount.getId(),
-                balance: newAccount.getBalance(),
-                owner_id: newAccount.getOwnerId(),
-                created_at: newAccount.getCreatedAt()
-            }
-    
-            await accountDatabase.insertAccount(newAccountDB)
-    
-            res.status(201).send(newAccount)
+            res.status(201).send(output)
         } catch (error) {
             console.log(error)
     
@@ -125,35 +78,17 @@ export class AccountController {
 
     public editAccountBalance = async (req: Request, res: Response) => {
         try {
-            const id = req.params.id
-            const value = req.body.value
-    
-            if (typeof value !== "number") {
-                res.status(400)
-                throw new Error("'value' deve ser number")
+
+            const input ={ 
+                id: req.params.id,
+                value: req.body.value 
             }
+
+            const accountBusiness = new AccountBusiness()
+            const output = await accountBusiness.editAccountBalance(input)
     
-            const accountDatabase = new AccountDatabase()
-            const accountDB = await accountDatabase.findAccountById(id)
-    
-            if (!accountDB) {
-                res.status(404)
-                throw new Error("'id' não encontrado")
-            }
-    
-            const account = new Account(
-                accountDB.id,
-                accountDB.balance,
-                accountDB.owner_id,
-                accountDB.created_at
-            )
-    
-            const newBalance = account.getBalance() + value
-            account.setBalance(newBalance)
-    
-            await accountDatabase.updateBalanceById(id, newBalance)
-    
-            res.status(200).send(account)
+            res.status(200).send(output)
+            
         } catch (error) {
             console.log(error)
