+++ src/database/arquitetura.sql
@@ -1,4 +1,4 @@
--- Active: 1675254981881@@127.0.0.1@3306
+-- Active: 1675686189180@@127.0.0.1@3306
 CREATE TABLE users (
     id TEXT PRIMARY KEY UNIQUE NOT NULL,
     name TEXT NOT NULL,
